Changes made after version 0.0.3 of a11y-auditor - fixing index.js and script-tag-specs.js
PACKAGE.JSON

Updated package.json with version 0.0.3. (Two versions published with
fixes after the 1st initial publish).

INDEX.JS:
If the config rules are passed as required (mentioned below),
function(“htmlSelector”, {
	‘selector_1’ : [‘array of rules to ignore’],
	‘selector_2’ : [‘array of rules to ignore’],
	‘selector_3’ : [‘*’] //skip all
}, false);
The normalizeConfig may not be needed (commented it for now).
Performed element.get(0). Modified config.ignore to config.
Even if the last parameter is left empty, things work & global rules
will not be executed as there is an undefined check wrapping the
section.

SCRIPT-TAG-SPECS.js
As of now, 3 tests fail :
1. Should be able to validate HTML string
2. Should be able to ignore one rules
3. Should not pollute the DOM

These are 3 negative test cases, so the check would have to be
something like expect().to.not.be.accessible() for it to pass -> but
this doesn’t work.